+++ aldrin/src/unknown_call.rs
@@ -44,7 +44,7 @@ impl UnknownCall {
 
     /// Deserializes the call's arguments into a generic [`Value`].
     pub fn deserialize_as_value(&self) -> Result<Value, DeserializeError> {
-        self.deserialize()
+        self.inner.as_ref().unwrap().deserialize_as_value()
     }
 
     /// Deserializes arguments and casts the call to a known [`Call`].