+++ aldrin/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `UnknownEvent`, which represents an unknown event.
 - Add variant `Error::InvalidEvent` and inner type `InvalidEvent`.
 - Add `Call` as a high-level alternative to `low_level::Call`.
+- Add `low_level::Call::deserialize_as_value()` to deserialize the arguments into a generic `Value`.
 
 ### Changed